add first mocha test
